test(e2e): drop API IDE coverage from the exploratory suite
The /apis/rest and /apis/graphql pages embed third-party IDEs (Swagger
UI, GraphiQL) whose behavior is not Phoenix's to test. Remove both
tests and rename the spec to exploratory-nav-chrome.spec.ts, which is
what the remaining tests actually cover (global search, side nav
persistence, breadcrumbs, account menu). Those tests now start from a
light page instead of loading an IDE.
