+++ js/app/tests/exploratory-playground.spec.ts
@@ -32,7 +32,10 @@ async function replaceCodeMirrorContent(
 }
 
 async function gotoPlayground(page: Page) {
-  await page.goto("/playground");
+  // The playground is the heaviest page in the app; wait for the document
+  // rather than every subresource so a fully parallel run doesn't exceed the
+  // navigation timeout. The assertions below still gate on real readiness.
+  await page.goto("/playground", { waitUntil: "domcontentloaded" });
   await page.waitForURL("**/playground");
   await expect(page.getByRole("heading", { name: "Playground" })).toBeVisible();
 }
